# Conflicts:
#	modules/runner/src/integrationTest/java/org/apache/ignite/internal/sql/api/ItSqlAsynchronousApiTest.java

+++ modules/runner/src/integrationTest/java/org/apache/ignite/internal/runner/app/client/ItThinClientSqlTest.java
@@ -517,7 +517,7 @@ public class ItThinClientSqlTest extends ItAbstractThinClientTest {
     }
 
     private static class Pojo {
-        public long num;
+        public int num;
 
         public String str;
     }
